Add Date/Time Tools page (#45)
* Add starting feature page

* Remove bad reference

* Add date-fns@^2.28.0

* Standardize text inputs

* Move global store

* Finish FieldIO prototype

* Reorganize some files

* Some minor code cleanup

* Move IO defs to class file

* Fix for weird reactivity issues

Primary issue: 'input' field was being emptied after `formatter.input` data was updated. Suspect cause was Vue seeing update to `formatter` object and re-rendering component that had bound `formatter.listener` prop.

Solution: instead of passing listener down to element and setting listener that way, set listener on wrapper component. `keyup` event is still captured correctly this way.

Secondary issue: `formatter.output` wasn't resolving correctly. Unsure of cause, but manually unwrapping ref value seems to fix. This is now done in a getter to mimic expected behavior.

* Make FieldIO class more reusable

* Use FieldIO class in text-tools

* Handful of important changes

1) Finish implementing parsing to unix timestamp
2) Simplify watching for radio button change event
3) Allow setting default radio button option
4) Update/fix grid layout issues
5) Update Text Tools radio button usage to match
6) Minor content updates

* showing to use v-model

* Add some piping tools

Initially I was going for the Pipeline and Pipe classes since I had a working implementation for that, but I decided on a new implementation specifically for functions.

* Support v-model on wrapper component

This extra event listener and emitter are needed because the `v-model` isn't otherwise added to the correct element.

* Add some piping tools

Initially I was going for the Pipeline and Pipe classes since I had a working implementation for that, but I decided on a new implementation specifically for functions.

* Rename to Frontend Stuff

Just rethinking having a side project (that's not a package/library) named after a framework.

* Remove old import of Nuxt global

* Set default handler directly

* Make TS happy

Wrap small chunk of logic in ridiculous, unnecessary `if` to appease the whiny type-checker.

* Bring back date-fns

* Fix up old imports and pipeline

* Fixup FieldIO class

* Start reformatting page

* Fix more type references and make component name consistent

* Install shadcn-vue + deps

* Configure shadcn-vue finally

* Types must be use type imports

* Start using some shadcn-vue components

* Input type selector works!

* rm debugging line

* prettier intensifies

* Reset default value

* Wrap Select components

* Add select for output format

* Fix defaults

* Bring back deleted value handler

Removing this broke existing features.

* Trying to make my Select be ok with its types

* Watch output selector value and update chosen filter

* 2 variations of ISO 8601 format

Still need error handling for something that can't be parsed into a Date

* minor spacing

* Stop crashing on parsing invalid date

* Add TBA labels to a couple of input options

* Add a human-readable option

* We have error messages!

* Remove some old imports

These were converted to Nuxt macros, so we no longer need to import their types.

* Update CHANGELOG

* Update CHANGELOG

---------

Co-authored-by: The Zach <zach2825@gmail.com>